+++ variant-calling/gatk/tests.sh
@@ -76,19 +76,22 @@ wgs_cmd_base="$WRAPPER -1 $WGS_FQ1 -2 $WGS_FQ2 -s NA12878-WGS -t WGS --name 'tes
 
 
 # DAG
-echo "DAG: WES" | tee -a $log
-odir=$(mktemp -d ${test_outdir_base}-wes.XXXXXXXXXX) && rmdir $odir
-eval $wes_cmd_base -o $odir -v --no-run >> $log 2>&1
-pushd $odir >> $log
-type=pdf;
-dag=example-dag.$type
-sed -i -e 's,num_chroms: .*,num_chroms: 1,' conf.yaml
-EXTRA_SNAKEMAKE_ARGS="--dag" bash run.sh; cat logs/snakemake.log | dot -T$type > $dag
-cp $dag $rootdir
-rm -rf $odir
-popd >> $log
+SKIP_DAG=1
+if [ $SKIP_DAG -eq 0 ]; then
+    echo "DAG: WES" | tee -a $log
+    odir=$(mktemp -d ${test_outdir_base}-wes.XXXXXXXXXX) && rmdir $odir
+    eval $wes_cmd_base -o $odir -v --no-run >> $log 2>&1
+    pushd $odir >> $log
+    type=pdf;
+    dag=example-dag.$type
+    sed -i -e 's,num_chroms: .*,num_chroms: 1,' conf.yaml
+    EXTRA_SNAKEMAKE_ARGS="--dag" bash run.sh; cat logs/snakemake.log | dot -T$type > $dag
+    cp $dag $rootdir
+    rm -rf $odir
+    popd >> $log
+fi
 
-    
+
 # dryruns
 #
 if [ $skip_dry_runs -ne 1 ]; then
